Ability to run tox in lxc environments
This is useful for snaps that we are testing that are using a different
core than default (for example, using core18 or core20). Users can now
pass in --environment 18.04 to lpmptox to run the tests entirely in
lxc.

Contents of your .ssh directory and .gitconfig (both in the $HOME
directory) will be copied into the lxc container.