Ensure any sudo calls also use the configured proxy
If a proxy has been configured on the host using environment
variables http_proxy or https_proxy then the container will inherit
this proxy config for the default user [1]. However then calling sudo
this proxy config is no longer configured. To fix this and so sudo
calls use the same proxy we can use `--preserve-env` [2] to pass
in a list of environment variables we want set for sudo user too.

We don't have to set proxy for pip install explicitly as if the http_proxy
or https_proxy variables are set then pip will use that proxy [3]. This is
the case for tox too [4].

[1] https://lxd.readthedocs.io/en/latest/environment/
[2] http://manpages.ubuntu.com/manpages/focal/man8/sudo.8.html
[3] https://pip.pypa.io/en/stable/user_guide/#using-a-proxy-server
[4] https://github.com/tox-dev/tox/pull/1500